refactor(runtime): isolate catalog owner resolution
