+++ src/cron/isolated-agent/run-model-selection.runtime.ts
@@ -2,11 +2,9 @@
 export { resolveAgentConfig } from "../../agents/agent-scope-config.js";
 export { resolveSubagentModelConfigSelectionResult } from "../../agents/agent-scope.js";
 export { DEFAULT_MODEL, DEFAULT_PROVIDER } from "../../agents/defaults.js";
-export {
-  loadResolvedPublishedModelCatalogOwner,
-  publishedModelCatalogOwnerMatchesAgent,
-  type ResolvedPublishedModelCatalogOwner,
-} from "../../agents/prepared-model-catalog.js";
+export { publishedModelCatalogOwnerMatchesAgent } from "../../agents/prepared-model-catalog-owner.js";
+export { loadResolvedPublishedModelCatalogOwner } from "../../agents/prepared-model-catalog.js";
+export type { ResolvedPublishedModelCatalogOwner } from "../../agents/prepared-model-catalog.types.js";
 export {
   getModelRefStatus,
   normalizeModelSelection,
